tests: change variable address's type to uint64_t

+++ tests/test.c
@@ -125,7 +125,7 @@ static void test()
 	};
 
 	csh handle;
-	size_t address = 0x1000;
+	uint64_t address = 0x1000;
 	//cs_insn insn[16];
 	cs_insn *insn;
 	int i;
